Added basic movement and jump
Added a basic movement and jump to the character.
Movement speed and jump height can be changed by editing the exported values in the JumpComponent and the MovementComponent on the Player Scene.
Added a script to parse JSON info to make animations quicker and save to the folder system. If there is an error with the animations delete the animations folder and it will remake them.
Added a basic state machine that will enter and exit states depending on inputs and the characters data.
Added a temporary dev spritesheet: https://mattwalkden.itch.io/free-space-runner-pack