Added a jump input forgiveness
Added an input forgiveness for jumping that will allow the Player to jump even after they've fallen off a ledge.
Added a JSON file to store player stats data and overrides the preset data in the MovementComponent and the JumpComponent. The data in the JSON now is the same as the default values.